requirements.txt
Django==3.1.5
django-cors-headers==3.7.0
django-decouple==2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0

